requirements.txt
click==8.0.3
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
passlib==1.7.4
psycopg2-binary==2.9.3
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.2
WTForms==3.0.1
